[RDRViews] added reasoner class, some doc update.
