+++ doc/views.md
@@ -1,0 +1,4 @@
+# Views
+
+## How to use the views
+
