+++ src/semantic_world/views/__init__.py
@@ -1,2 +1,1 @@
-from .world_rdr import *
 from .views import *
